충돌판정 완성 게임종료 표시 시간,미사일수 표시

+++ Engine/닷지/Engine.cpp
@@ -4,8 +4,10 @@
 namespace Rendering
 {
 	void initialize(HWND const& hwindow);
-	void update(HWND const& hwindow, int player_x, int player_y,
-		std::vector<POINT> const& missilepos);
+	void update(HWND const& hwindow, int const player_x, int const player_y,
+		std::vector<POINT> const& missilepos, float const time, bool const is_printover);
+	//void gameover(HWND const& hwindow, int const player_x, int const player_y,
+	//	std::vector<POINT> const& missilepos);
 	void destroy();
 }
 
@@ -16,13 +18,16 @@ namespace Input
 	void initplayerpos();
 	POINT const playerpos();
 	void input();
+	bool enter();
 }
 namespace Time
 {
 	void Procedure
 	(HWND const , UINT const , WPARAM const , LPARAM const );
-	bool const isinterval();
-	bool const isregentime();
+	bool isinterval();
+	bool isregentime();
+	bool isprinttime();
+	float getdelta();
 }
 namespace Missile
 {
@@ -30,12 +35,21 @@ namespace Missile
 	void move_missile();
 	void delete_missile();
 	std::vector<POINT> const vecpoint();
+	size_t getcount();
+}
+namespace Collision
+{
+	bool Procedure
+	(HWND   const hWindow, UINT   const uMessage, WPARAM const wParameter,
+		LPARAM const lParameter, std::vector<POINT> const& missile, POINT const& player);
+	bool iscrashed(std::vector<POINT> const& missile,
+		POINT const& player);
 }
 
 namespace Engine
 {
-
-
+	bool isalive;
+	float record_time;
 	LRESULT CALLBACK Procedure
 	(HWND hwindow, UINT umessage, WPARAM wparameter, LPARAM lparameter)
 	{
@@ -45,22 +59,41 @@ namespace Engine
 			{
 				Rendering::initialize(hwindow);
 				Input::initplayerpos();
+				isalive =true;
+				record_time=0;
 				return 0;
 			}
 			case WM_APP:
 			{
 				Time::Procedure(hwindow, umessage, wparameter, lparameter);
-				
-				if(Time::isregentime())
-					Missile::push_missile(Input::playerpos());
-				if(Time::isinterval())
+				if(isalive)
 				{
-					Input::input();
-					Missile::move_missile();
-					Missile::delete_missile();
+					
+					if(Time::isregentime())
+						Missile::push_missile(Input::playerpos());
+					if(Time::isinterval())
+					{
+						Input::input();
+						Missile::move_missile();
+						Missile::delete_missile();
+						isalive = !Collision::iscrashed(Missile::vecpoint(),Input::playerpos());
+					}
+					record_time+=Time::getdelta();
+					Rendering::update(hwindow, Input::playerpos().x, Input::playerpos().y,
+						Missile::vecpoint(), record_time, false);
 				}
-				Rendering::update(hwindow, Input::playerpos().x, Input::playerpos().y,
-					Missile::vecpoint());
+				else
+				{
+					if(Time::isprinttime())
+						Rendering::update(hwindow, Input::playerpos().x, Input::playerpos().y,
+							Missile::vecpoint(),record_time,true);
+					else
+						Rendering::update(hwindow, Input::playerpos().x, Input::playerpos().y,
+							Missile::vecpoint(), record_time, false);
+
+
+				}
+	
 
 				return 0;
 			}
@@ -77,7 +110,6 @@ namespace Engine
 
 			case WM_DESTROY:
 			{
-
 				Rendering::destroy();
 				ExitProcess(0);
 				return 0;
